Move test cases into own file
